Temporary changes, TODO: Move converted to Nox.Solution project Fix validation schema
